+++ go/iden3mobile/config-example.yml
