+++ go/iden3mobile/identity.go
@@ -5,22 +5,23 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"strings"
 	"time"
 
-	"github.com/ethereum/go-ethereum/accounts"
-	ethkeystore "github.com/ethereum/go-ethereum/accounts/keystore"
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/ethclient"
+	"github.com/google/uuid"
+	"github.com/iden3/go-iden3-core/components/httpclient"
 	"github.com/iden3/go-iden3-core/components/idenpuboffchain/readerhttp"
-	"github.com/iden3/go-iden3-core/components/idenpubonchain"
 	"github.com/iden3/go-iden3-core/db"
-	"github.com/iden3/go-iden3-core/eth"
 	"github.com/iden3/go-iden3-core/identity/holder"
-	babykeystore "github.com/iden3/go-iden3-core/keystore"
 	"github.com/iden3/go-iden3-crypto/babyjub"
+	issuerMsg "github.com/iden3/go-iden3-servers-demo/servers/issuerdemo/messages"
+	verifierMsg "github.com/iden3/go-iden3-servers-demo/servers/verifier/messages"
 	log "github.com/sirupsen/logrus"
 )
+
+type Callback interface {
+	VerifierResHandler(bool, error)
+	RequestClaimResHandler(*Ticket, error)
+}
 
 type Identity struct {
 	id          *holder.Holder
@@ -36,7 +37,22 @@ const (
 	storageSubPath       = "/idStore"
 	keyStorageSubPath    = "/idKeyStore"
 	smartContractAddress = "0xF6a014Ac66bcdc1BF51ac0fa68DF3f17f4b3e574"
+	credExisPrefix       = "credExis"
 )
+
+func isEmpty(path string) (bool, error) {
+	f, err := os.Open(path)
+	if err != nil {
+		return false, err
+	}
+	defer f.Close()
+
+	_, err = f.Readdirnames(1)
+	if err == io.EOF {
+		return true, nil
+	}
+	return false, err
+}
 
 // NewIdentity creates a new identity
 // this funciton is mapped as a constructor in Java.
@@ -158,84 +174,63 @@ func (i *Identity) Stop() {
 	i.stopTickets <- true
 }
 
-func loadComponents(storePath, web3Url string) (idenpubonchain.IdenPubOnChainer, *babykeystore.KeyStore, db.Storage, error) {
-	iPub, err := loadIdenPubOnChain(web3Url)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	storage, err := loadStorage(storePath)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	ks, err := loadKeyStoreBabyJub(storePath)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	return iPub, ks, storage, nil
+// RequestClaim sends a petition to issue a claim to an issuer.
+// This function will eventually trigger an event,
+// the returned ticket can be used to reference the event
+func (i *Identity) RequestClaim(baseUrl, data string, c Callback) {
+	go func() {
+		id := uuid.New().String()
+		t := &Ticket{
+			Id:     id,
+			Type:   TicketTypeClaimStatus,
+			Status: TicketStatusPending,
+		}
+		httpClient := httpclient.NewHttpClient(baseUrl)
+		res := issuerMsg.ResClaimRequest{}
+		if err := httpClient.DoRequest(httpClient.NewRequest().Path(
+			"claim/request").Post("").BodyJSON(&issuerMsg.ReqClaimRequest{
+			Value: data,
+		}), &res); err != nil {
+			c.RequestClaimResHandler(nil, err)
+			return
+		}
+		t.handler = &reqClaimStatusHandler{
+			Id:      res.Id,
+			BaseUrl: baseUrl,
+		}
+		err := i.Tickets.Add([]Ticket{*t})
+		c.RequestClaimResHandler(t, err)
+	}()
 }
 
-func loadStorage(baseStorePath string) (db.Storage, error) {
-	storagePath := baseStorePath + storageSubPath
-	// Open database
-	storage, err := db.NewLevelDbStorage(storagePath, false)
-	if err != nil {
-		return nil, fmt.Errorf("Error opening leveldb storage: %w", err)
-	}
-	log.WithField("path", storagePath).Info("Storage opened")
-	return storage, nil
-}
-
-func loadKeyStoreBabyJub(baseStorePath string) (*babykeystore.KeyStore, error) {
-	storagePath := baseStorePath + keyStorageSubPath
-	// Open keystore
-	storage := babykeystore.NewFileStorage(storagePath)
-	ks, err := babykeystore.NewKeyStore(storage, babykeystore.StandardKeyStoreParams)
-	if err != nil {
-		return nil, fmt.Errorf("Error creating/opening babyjub keystore: %w", err)
-	}
-	return ks, nil
-}
-
-// WARNING: THIS CODE IS COPIED FROM go-iden3-servers/loaders/loaders.go
-func loadEthClient2(ks *ethkeystore.KeyStore, acc *accounts.Account, web3Url string) (*eth.Client2, error) {
-	// TODO: Handle the hidden: thing with a custon configuration type
-	hidden := strings.HasPrefix(web3Url, "hidden:")
-	if hidden {
-		web3Url = web3Url[len("hidden:"):]
-	}
-	client, err := ethclient.Dial(web3Url)
-	if err != nil {
-		return nil, fmt.Errorf("Error dialing with ethclient: %w", err)
-	}
-	if hidden {
-		log.WithField("url", "(hidden)").Info("Connection to web3 server opened")
-	} else {
-		log.WithField("url", web3Url).Info("Connection to web3 server opened")
-	}
-	return eth.NewClient2(client, acc, ks), nil
-}
-
-func loadIdenPubOnChain(web3Url string) (idenpubonchain.IdenPubOnChainer, error) {
-	client, err := loadEthClient2(nil, nil, web3Url)
-	if err != nil {
-		return nil, err
-	}
-	addresses := idenpubonchain.ContractAddresses{
-		IdenStates: common.HexToAddress(smartContractAddress), // TODO: hardcode the address
-	}
-	return idenpubonchain.New(client, addresses), nil
-}
-
-func isEmpty(path string) (bool, error) {
-	f, err := os.Open(path)
-	if err != nil {
-		return false, err
-	}
-	defer f.Close()
-
-	_, err = f.Readdirnames(1)
-	if err == io.EOF {
-		return true, nil
-	}
-	return false, err
+// ProveCredential sends a credentialValidity build from the given credentialExistance to a verifier
+// the callback is used to check if the verifier has accepted the credential as valid
+func (i *Identity) ProveClaim(baseUrl string, credId []byte, c Callback) {
+	// TODO: add context
+	go func() {
+		// Get credential existance
+		credExis, err := i.ClaimDB.GetReceivedCredential(credId)
+		if err != nil {
+			c.VerifierResHandler(false, err)
+			return
+		}
+		// Build credential validity
+		credVal, err := i.id.HolderGetCredentialValidity(credExis)
+		if err != nil {
+			c.VerifierResHandler(false, err)
+			return
+		}
+		// Send credential to verifier
+		httpClient := httpclient.NewHttpClient(baseUrl)
+		if err := httpClient.DoRequest(httpClient.NewRequest().Path(
+			"verify").Post("").BodyJSON(verifierMsg.ReqVerify{
+			CredentialValidity: credVal,
+		}), nil); err != nil {
+			// Credential declined / error
+			c.VerifierResHandler(false, err)
+			return
+		}
+		// Success
+		c.VerifierResHandler(true, nil)
+	}()
 }
