+++ go/iden3mobile/identity_test.go
