+++ go/iden3mobile/mobilearrays.go
